tree: Move structures to private.h
Move generic structures to private.h to avoid having to duplicate
definitions in the SWIG interface file.
