+++ src/nvme/json.c
@@ -13,7 +13,6 @@
 #include <json-c/json.h>
 
 #include "fabrics.h"
-#include "private.h"
 
 #define json_object_add_value_string(o, k, v)			\
 	json_object_object_add(o, k, json_object_new_string(v))
